TaskChecker/requirements.txt
aiokafka==0.8.1
anyio==3.7.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
docker==6.1.3
epicbox==1.1.0
exceptiongroup==1.1.1
fastapi==0.98.0
idna==3.4
kafka-python==2.0.2
packaging==23.1
pycodestyle==2.10.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
structlog==23.1.0
typing_extensions==4.6.3
urllib3==1.26.14
websocket-client==1.6.0
